Dynamically load DB implementation error checking.
When trying to set up e2e tests, the cgo dependency that go-sqlite3
introduced caused issues when trying to run ko built containers. While
enabling CGO_ENABLED=1 would allow the image to built by ko, trying to
run it would encounter the error:

```
cgo enabled standard_init_linux.go:219: exec user process caused: no such file or directory
```

Most threads suggested disabling CGO_ENABLED to resolve this.

Instead of trying to fight with runtimes to get the CGO image to work,
since we only use sqlite3 for local unit testing this change modifies
the error handling to dynamically load the error space for DB
implementation specific error handling. This allows us to isolate usage
of cgo to only unit tests, removing the dependency from the ko built
image.
